moved to ObjectStorage trait
